refactor(fieldmap): move utils.ts to utils/functions.ts
